Port 4.3 Theme Engine to 4.4 (WIP)
Change-Id: Ic6038a8c2d6b3e03e825e335db877d83620c8e0e

Theme Engine: fixing for 4.4 compatibility

Change-Id: Ie27e13b27a4f6ba390112ec09b35598d3b085e17

Theme Engine: let theme service handle saving icons

Change-Id: Iaeef72242d9a043b6da62cb1f0ff2081058f556b

Boot ani: Allow background color to be specified

An RGB value may be added to the end of the first line
to specify a color to be used as the background instead
of black.  Black is the default when no color is
specified.

PS2:
Preserve custom background color when scaling animation
in ThemeManagerService.

Change-Id: Icf242a584513cfe974ef42f21f37a5d10400518b

update

Change-Id: Ib59365497bc79c65736f16b31f206994ac667428