[systemUI] fix rounding issues and take density into account
we are a multiDpi rom...and on 4.2 and 4.3 already the text
in the circle battery icon was on lower dpi a small bit up. The original
author compensated this with just a +1 to the Y value. Thats not the right
way...calculate the current dpi into the offset and all is well.

We want a perfect 4.4...so this nitpicks even if we are talking about
0.5dp maybe should be fixed

Change-Id: I8e12399ad12a2b5e4b7e69c3de2b63fbb6ae18e6