Add navbar customizations [1/2]
This patch adds:
- Navbar up to 5 buttons
- dynamic menu behaviour
- landscape navbar options as well now on tablets
- button glow color
- button icon color
- navbar user can set system or custom icons
- navbar height and width
- enable/disable navbar
- actions are vor navbar and available for the whole system over MagicActions.java
	back
	custom app or activtiy
	home
	kill app
	last app
	notification drawer
	ime switcher
	power menu
	qs pannel
	recents screen
	screen off
	screenshot
	search
	torch if it is available on the device

===============================
Fix: external storage problem NavRing and NavBar applications

If the user binds an app which is located on the
external sdcard and is rebooting the button icon
will be left blank because the navring/navbar is already
constructed before the sdcard is mounted. To avoid
this we listen to the app available intents.
If it is not available the ring target/navbar button shows the none icon.
 When available it shows the
correct app icon. So that the user can clearly see
this app is accessable or not. As well when the
user goes into mass storage mode or does
unmount the sdcard or whatever case could
happen that the application is not there.

==================================
enable navbar for hw key devices

we moved this part into PhoneWindowManager which controlls the navbar.
Due that we allow on all devices a navbar the check on PhoneStatusBar
is not needed and prevend the navbar on comming up on hw key devices

==================================
 do not inflate twice

small derp does not effect anything but as we know inflating is slow

Ported from slim rom

Change-Id: I2622e917d875460b941c8b56e2f8321f5a78ea57