[SystemUI] Squash commits of quick settings [1/2]
Squash of quick settings from slimrom

Include:

https://github.com/SlimRoms/frameworks_base/commit/d85304fac9bde550beeeeca0dba2cd49686f20d7

Quick Settings Pulldown
https://github.com/SlimRoms/frameworks_base/commit/cfc1eb63afab2df03b720ea375cb9ee93298745b

    fb: qs tweeks

    - fix fc problem on tablets
    - fix filtering of unsupported tiles
    - add music tile options (disable artwork, disable track title optional)
    - add instance handling as well to the tile chooser dialog
    - add cancel button to chooser dialog
    - battery tile short press now battery stats - longpress batteryicon settings
    - rotation tile lonpress - rotation settings
    - collapse drawer on go to sleep tile

https://github.com/SlimRoms/frameworks_base/commit/4bc584e721c9558657deca7ac9de31b7f0ca3af3

QS: make dynamic tiles configurable (1/2)

instead of using a bunch of provider settings var
like it was on 4.3 and almost all other roms are doing
use one and split like we do for the normal tiles. This
is more consistent and saves us to observe a bunch of values
As well this is integrated into the one layout screen we have. So
all QS related is easy and logical usable for the user on one
screen
https://github.com/SlimRoms/frameworks_base/commit/8a25648e21513382018c1d465ab902d1778d2d98

SystemUI: add CHANGE_NETWORK_MODE permission

Required for 2G/3G quick settings toggle

fb: fix qs battery tile on normal icon due of left over

I forgot to commit a change so on icon and icon percentage there
was still a margin set. Remove the magin to fix it

fb: allow to disable quicksettings

on 4.3 we had an extra settings option for it. Due that we want
to reduce the settings to improve the UX and due that
we are doing it this way on all dslv settings we change
the old 4.3 way.

Now it is very easy like dslv.....remove all tiles and QS
is completly disabled and cannot be opend any more. Notification
drawer shows the settings icon on the top right in this case.
As well compared to 4.3 this works on phones and tablets

Change-Id: I4cebc43e41cb5f54d17a4e8b3c94bb2a79d0879c
TODO: add a hint for the user in settings when disabled like in dslv

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBar.java

QS: music tile update on Uri change

gives the user a bit better feeling when changing
the tile settings in Settings due it does react
immediately

Change-Id: I162477bb65eab7e0283aae135aeb013f99e38abc

QS: enable haptic feedback and change Alarm tile longpress

nice idea from xploid to have haptic feedback on the tiles
just a bit to complicated to do on every tile

so activate it centralized

as well change the longressbehaviour of alarm tile to the
normal onClick behaviour

Change-Id: I4cebc43e41cb5f54d17a4e8b3c94bb2a79d0879c